Extract CacheBufferedIndexInput
